requirements.txt
ccxt
pandas
matplotlib
streamlit
python-dotenv